test(cli): split readonly-commands into per-domain files on the TTY harness
readonly-commands.test.ts (1,443 lines / 45 tests) split along the B5.3
command-module seams into seven files: help-errors, plugins,
tasks-workflows, agents-packages, settings-paths, search-trash-schedule,
and logs. Two mega-tests whose segments were self-contained (same mocks,
independent asserts) were de-chained (45 → 48 tests); the taxonomy --json
test stays intact because its fetch-call indexes are order-dependent.
Assertions are unchanged (1,301 expect() calls before and after). The
logs file keeps the real content-dir + temp BAKIN_HOME pattern via
withTempBakinHome; the rest carry the defensive dual content-dir mocks.
